requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
braceexpand==0.1.7
click==8.1.7
discord.py @ git+https://github.com/DA-344/d.py.git@c0ca0fdd8baa2b2f9f2f7eefb7065ee265f483b8
frozenlist==1.4.1
idna==3.6
import-expression==1.1.4
jishaku==2.5.2
multidict==6.0.5
six==1.16.0
wheel==0.43.0
yarl==1.9.4
